+++ src/index.ts
@@ -2,3 +2,5 @@ export * from './format-date-time/format-date-time';
 export * from './date-time-difference/date-time-difference';
 export * from './index-to-name/index-to-name';
 export * from './add-subtract-time/add-subtract-time';
+export * from './is-valid-date/is-valid-date';
+export * from './is-leap-year/is-leap-year';